java-examples-wiremock: WebClient - Delete a movie HTTP DELETE

+++ java-unittests/WireMock/wiremock-movie-app/src/main/java/com/fral/wiremock/service/MoviesRestClient.java
@@ -167,7 +167,7 @@ public class MoviesRestClient {
         return updatedMovie;
     }
 
-    public String deleteMovieById(Integer movieId) {
+    public String deleteMovieById(Long movieId) {
 
        String response;
         try {
